Apply the deprecation prescriptions: 133 sites across eleven families
A thirteen-agent analysis pass produced exact, evidence-backed prescriptions for every remaining
deprecated call outside Secure Transport; this commit applies them. The families: the toolbar
separator and customize identifiers, no-ops since 10.7, removed from four toolbars; all fifteen
OSAtomic sites in the transcript indexer moved to stdatomic with the five variables retyped
_Atomic, so every plain read keeps its full-barrier meaning; the three alert sheets in the
transcript viewer on the completion-handler form; seventeen drawing-path calls, from context:nil
through the deletion of lockFocus pairs inside private drop-highlight hooks; sixteen constant and
system-color renames, where the analysis corrected its own briefing twice (NSTokenStyleNone, not a
name that does not exist, and NSBezelStyleFlexiblePush all the way through); the old collection
view grid API deliberately kept behind measured pragmas, because both xibs run the cell-based
engine and a grid layout provably kills item creation; NSArchiver replaced by a keyed archive on
the write side and a three-format fallback reader, so no stored preference is eaten; the
NSConnection vend deleted after a repo-wide proof that nothing ever connected to it; the workspace
and LaunchServices lookups modernised with UniformTypeIdentifiers linked into both projects; the
image-picker drag rebuilt on NSFilePromiseProvider; the uploader and the Xtras downloader rewritten
on NSURLSession, which incidentally made the uploader's thirty-second timeout real for the first
time, since its timer had never been scheduled; and the last WebKit-1 reference in the Adium
framework type-erased, so the framework no longer touches WebKit at all, closing on the way a hole
where paste-and-match-style fetched remote images.

Two real bugs surfaced by side effect: the downloader's progress bar compared an unsigned size to
-1 and so never had a maximum, and the avatar-scaling flip flag was measured end to end and found
inert for every data-backed image, so it is gone.

This commit is one half of a pair applied to the same verified tree; its sibling converts 113
files to ARC. Each half's files are disjoint, the tree with both builds clean with 68 deprecated
calls left (from 175), and the two halves were not build-tested in isolation.

+++ Other/XtrasCreator/NSMutableArrayAdditions.h
@@ -10,6 +10,6 @@
 @interface NSMutableArray (NSMutableArrayAdditions)
 
 //assumes a sorted array. does the Right Thing for empty arrays.
-- (unsigned) indexForInsortingObject:(id)obj usingSelector:(SEL)compareCmd;
+- (NSUInteger) indexForInsortingObject:(id)obj usingSelector:(SEL)compareCmd;
 
 @end
